Add files via upload
Description About Rock Paper Scissor: 
This is an interactive Rock–Paper–Scissors game built using HTML, CSS, and JavaScript. The user selects Rock, Paper, or Scissors, and the computer randomly chooses a move. The game updates scores in real-time and displays messages showing the winner of each round. It provides a simple and fun way to practice basic JavaScript logic and DOM manipulation. 

Description About Tic Tac Toe : 
This is a simple Tic-Tac-Toe game built using HTML, CSS, and JavaScript. Two players take turns marking X and O on a 3×3 grid, and the game detects wins, draws, and resets. It provides an easy and fun way to practice JavaScript logic and DOM manipulation.